Fix auth tests. Refactor authentication tests to use promises and integrate with ctx.
